# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso."
          },
          "400": {
            "description": "Email já está em uso."
          }
        },
        "summary": "Registrar novo usuário com senha",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "401": {
            "description": "Credenciais inválidas ou usuário bloqueado."
          }
        },
        "summary": "Autenticar usuário com e-mail e senha",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens renovados com sucesso."
          },
          "401": {
            "description": "Refresh token inválido ou usuário bloqueado."
          }
        },
        "summary": "Renovar tokens de acesso usando o refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Código enviado com sucesso."
          },
          "429": {
            "description": "Aguarde antes de solicitar novo código."
          }
        },
        "summary": "Solicitar código OTP de 6 dígitos por e-mail",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada e tokens retornados."
          },
          "400": {
            "description": "Código OTP incorreto ou expirado."
          },
          "429": {
            "description": "Limite de tentativas excedido."
          }
        },
        "summary": "Validar código OTP de 6 dígitos e iniciar sessão (Passwordless / Auth)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão encerrada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revogar sessão atual (Server-side logout)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Todas as sessões encerradas."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revogar todas as sessões ativas do usuário",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password/forgot": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitação processada."
          }
        },
        "summary": "Solicitar código OTP para redefinição de senha",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha redefinida com sucesso."
          },
          "400": {
            "description": "Código inválido ou expirado."
          }
        },
        "summary": "Redefinir senha utilizando o código OTP de verificação",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/admin/dashboard/overview": {
      "get": {
        "operationId": "AnalyticsController_getOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resumo geral da plataforma (KPIs)",
        "tags": [
          "Admin - Dashboard",
          "Admin - Analytics"
        ]
      }
    },
    "/admin/dashboard/revenue": {
      "get": {
        "operationId": "AnalyticsController_getRevenue",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Analytics Financeiro / Faturamento",
        "tags": [
          "Admin - Dashboard",
          "Admin - Analytics"
        ]
      }
    },
    "/admin/dashboard/ai-usage": {
      "get": {
        "operationId": "AnalyticsController_getAiUsage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Analytics de Inteligência Artificial",
        "tags": [
          "Admin - Dashboard",
          "Admin - Analytics"
        ]
      }
    },
    "/admin/dashboard/top-destinations": {
      "get": {
        "operationId": "AnalyticsController_getTopDestinations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Destinos mais populares",
        "tags": [
          "Admin - Dashboard",
          "Admin - Analytics"
        ]
      }
    },
    "/admin/dashboard/users-growth": {
      "get": {
        "operationId": "AnalyticsController_getUsersGrowth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crescimento de base de usuários",
        "tags": [
          "Admin - Dashboard",
          "Admin - Analytics"
        ]
      }
    },
    "/admin/dashboard/trips-growth": {
      "get": {
        "operationId": "AnalyticsController_getTripsGrowth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crescimento de roteiros gerados",
        "tags": [
          "Admin - Dashboard",
          "Admin - Analytics"
        ]
      }
    },
    "/admin/dashboard/storage": {
      "get": {
        "operationId": "AnalyticsController_getStorageStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Monitoramento de mídia estática",
        "tags": [
          "Admin - Dashboard",
          "Admin - Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "2GO / AppRoteiros API",
    "description": "API de Roteiros de Viagem com IA e Autenticação OTP",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "password": {
            "type": "string",
            "example": "senhaForte123",
            "minLength": 6
          }
        },
        "required": [
          "fullName",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "password": {
            "type": "string",
            "example": "senhaForte123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "seu-refresh-token-aqui"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "usuario@email.com",
            "description": "E-mail do usuário para recebimento do código de 6 dígitos"
          },
          "purpose": {
            "type": "string",
            "example": "LOGIN",
            "enum": [
              "SIGNUP",
              "LOGIN",
              "EMAIL_VERIFICATION",
              "PASSWORD_RESET"
            ],
            "description": "Propósito do código OTP (SIGNUP, LOGIN, EMAIL_VERIFICATION, PASSWORD_RESET)"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "usuario@email.com",
            "description": "E-mail do usuário"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Código OTP de 6 dígitos recebido por e-mail"
          },
          "purpose": {
            "type": "string",
            "example": "LOGIN",
            "enum": [
              "SIGNUP",
              "LOGIN",
              "EMAIL_VERIFICATION",
              "PASSWORD_RESET"
            ],
            "description": "Propósito do código OTP (SIGNUP, LOGIN, EMAIL_VERIFICATION, PASSWORD_RESET)"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "usuario@email.com",
            "description": "E-mail cadastrado do usuário para recuperação de senha"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "usuario@email.com",
            "description": "E-mail cadastrado do usuário"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Código OTP de 6 dígitos recebido para redefinição de senha"
          },
          "newPassword": {
            "type": "string",
            "example": "NovaSenhaSegura123!",
            "description": "Nova senha do usuário (mínimo de 6 caracteres)"
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
...
    }
  }
}